+++ src/pages/SignUp.tsx
@@ -1,14 +1,12 @@
-import { useNavigate } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
 
-import { signUp } from 'api/auth.api';
-import { useAlert } from 'hooks/useAlert';
 import { IAuthData } from 'models/user.model';
 import { EmailFieldset, PasswordFieldset, AuthForm } from 'components/AuthForm';
+import { useAuth } from 'hooks/useAuth';
 
 const SignUp = () => {
-  const navigate = useNavigate();
-  const { showAlert } = useAlert();
+  const { userSignUp } = useAuth();
+
   const {
     register,
     handleSubmit,
@@ -16,15 +14,7 @@ const SignUp = () => {
   } = useForm<IAuthData>();
 
   const onSubmit = handleSubmit((data: IAuthData) => {
-    signUp(data)
-      .then((res) => {
-        showAlert('회원가입이 완료되었습니다.');
-        navigate('/login');
-      })
-      .catch((error) => {
-        const errorMsg = error.response.data[0].msg ?? '회원가입에 실패했습니다. 다시 시도 해주세요.';
-        showAlert(errorMsg);
-      });
+    userSignUp(data);
   });
 
   return (
